+++ microsoft_file_share_edm/main.py
@@ -27,6 +27,7 @@ from .utils.constants import (
     PLUGIN_NAME,
     PLUGIN_VERSION,
     MODULE_NAME,
+    SAMPLE_CSV_FILE_NAME,
 )
 from netskope.integrations.edm.plugin_base import PluginBase, ValidationResult
 from netskope.integrations.edm.utils import CONFIG_TEMPLATE, FILE_PATH
@@ -233,7 +234,15 @@ class MicrosoftFileShareEDMPlugin(PluginBase):
             )
 
             # Validate that the file exists and is accessible on the remote server
-            protocol_object.validate_remote_file(server_configuration)
+            try:
+                input_path = f"{FILE_PATH}/{self.name}/{SAMPLE_CSV_FILE_NAME}"
+                self.storage.update({"csv_path": input_path})
+                self.pull_sample_data()
+            except Exception as error:
+                return ValidationResult(
+                    success=False,
+                    message=str(error),
+                )
 
             self.logger.debug(
                 (
@@ -278,6 +287,26 @@ class MicrosoftFileShareEDMPlugin(PluginBase):
                 success=True,
                 message="Step validated successfully.",
             )
+            input_path = f"{FILE_PATH}/{self.name}/{SAMPLE_CSV_FILE_NAME}"
+            output_path = f"{FILE_PATH}/{self.name}"
+            try:
+                self.storage.update({
+                    "csv_path": input_path,
+                    "sanitization_data_path": output_path
+                })
+                self.sanitize(file_name="sample", sample_data=True)
+            except Exception:
+                error_message = (
+                    f"Error occurred while sanitizing the sample data for configuration '{self.name}'."
+                )
+                self.logger.error(
+                    message=f"{self.log_prefix} {error_message}",
+                    details=traceback.format_exc(),
+                )
+                result = ValidationResult(
+                    success=False,
+                    message=error_message,
+                )
         elif step == "sanity_results":
             result = ValidationResult(
                 success=True,
@@ -586,19 +615,12 @@ class MicrosoftFileShareEDMPlugin(PluginBase):
                 if name == "sanity_inputs":
                     for field in fields:
                         if field["type"] == "sanitization_input":
-                            input_path = f"{FILE_PATH}/{self.name}/sample.csv"
+                            input_path = f"{FILE_PATH}/{self.name}/{SAMPLE_CSV_FILE_NAME}"
                             self.storage.update({"csv_path": input_path})
                             field["default"] = self.pull_sample_data()
                 elif name == "sanity_results":
                     for field in fields:
                         if field["type"] == "sanitization_preview":
-                            input_path = f"{FILE_PATH}/{self.name}/sample.csv"
-                            output_path = f"{FILE_PATH}/{self.name}"
-                            self.storage.update({
-                                "csv_path": input_path,
-                                "sanitization_data_path": output_path
-                            })
-                            self.sanitize(file_name="sample", sample_data=True)
                             field["default"] = {
                                 "sanitizationStatus": True,
                                 "message": "Sanitization Done Successfully",
@@ -1206,6 +1228,13 @@ class SMBProtocolFileShareEDMPlugin:
 
                 except MicrosoftFileShareEDM:
                     raise
+                except UnicodeDecodeError as error:
+                    error_message = (
+                        "Only CSV file is supported. Please provide a valid CSV file."
+                    )
+                    raise MicrosoftFileShareEDM(
+                        value=error, message=error_message
+                    ) from error
                 except Exception as error:
                     error_message = f"File '{remote_file_path}' exists but is not readable or not a valid CSV file."
                     raise MicrosoftFileShareEDM(
@@ -1513,6 +1542,11 @@ class SFTPProtocolFileShareEDMPlugin:
                                 error_message = f"File '{remote_file_path}' is empty or not readable as CSV."
                                 raise MicrosoftFileShareEDM(message=error_message)
 
+                    except UnicodeDecodeError as error:
+                        error_message = "Only CSV file is supported. Please provide a valid CSV file."
+                        raise MicrosoftFileShareEDM(
+                            value=error, message=error_message
+                        ) from error
                     except PermissionError as error:
                         error_message = f"File '{remote_file_path}' exists but is not readable on the remote server."
                         raise MicrosoftFileShareEDM(