+++ microsoft_file_share_edm/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.1
+## Fixed
+- Fixed plugin validation with respect to CSV path.
+- Fixed Sanity Step level validation.
+
 # 1.0.0
 ## Added
 - Initial release.